minor fixes

+++ test/test_cnf_simplifiy.c
@@ -28,7 +28,7 @@ test_Cnf_simplify_withEmptyAssignment()
 
   assert(cnf.count == resultCnf.count);
 
-  for (int i = 0; i < cnf.count; ++i) {
+  for (size_t i = 0u; i < cnf.count; ++i) {
     assert(cnf.pData[i] == resultCnf.pData[i]);
   }
 }
@@ -38,10 +38,10 @@ test_Cnf_simplify_withTrueClause()
 {
   Assignment assignment;
   Assignment_create(&assignment);
-  Assignment_set(&assignment, 1, 1);
+  Assignment_set(&assignment, 1u, 1);
 
   int32_t clause[] = { 1, -2, 3 };
-  size_t clauseCount = 3;
+  size_t clauseCount = 3u;
 
   Cnf cnf;
   Cnf_create(&cnf);
@@ -52,7 +52,7 @@ test_Cnf_simplify_withTrueClause()
 
   Cnf_simplify(&cnf, &assignment, &resultCnf);
 
-  assert(resultCnf.count == 0);
+  assert(resultCnf.count == 0u);
 }
 
 void
@@ -60,10 +60,10 @@ test_Cnf_simplify_withFalseClause()
 {
   Assignment assignment;
   Assignment_create(&assignment);
-  Assignment_set(&assignment, 1, 0);
+  Assignment_set(&assignment, 1u, 0);
 
   int32_t clause[] = { 1, -2, 3 };
-  size_t clause1Count = 3;
+  size_t clause1Count = 3u;
 
   Cnf cnf;
   Cnf_create(&cnf);
@@ -74,7 +74,7 @@ test_Cnf_simplify_withFalseClause()
 
   Cnf_simplify(&cnf, &assignment, &resultCnf);
 
-  assert(resultCnf.count == (cnf.count - 1));
+  assert(resultCnf.count == (cnf.count - 1u));
 }
 
 void
@@ -82,16 +82,16 @@ test_Cnf_simplify_withMixedClauses()
 {
   Assignment assignment;
   Assignment_create(&assignment);
-  Assignment_set(&assignment, 1, 0);
-  Assignment_set(&assignment, 5, 1);
+  Assignment_set(&assignment, 1u, 0);
+  Assignment_set(&assignment, 5u, 1);
 
   // this clause should have the 1 omitted
   int32_t clause1[] = { 1, -2, 3 };
-  size_t clause1Count = 3;
+  size_t clause1Count = 3u;
 
   // this clause should be absent from the result
   int32_t clause2[] = { 3, -4, 5 };
-  size_t clause2Count = 3;
+  size_t clause2Count = 3u;
 
   Cnf cnf;
   Cnf_create(&cnf);
@@ -103,11 +103,11 @@ test_Cnf_simplify_withMixedClauses()
 
   Cnf_simplify(&cnf, &assignment, &resultCnf);
 
-  assert(resultCnf.count == 4);
-  assert(resultCnf.pData[0] == 0);
-  assert(resultCnf.pData[1] == -2);
-  assert(resultCnf.pData[2] == 3);
-  assert(resultCnf.pData[0] == 0);
+  assert(resultCnf.count == 4u);
+  assert(resultCnf.pData[0u] == 0);
+  assert(resultCnf.pData[1u] == -2);
+  assert(resultCnf.pData[2u] == 3);
+  assert(resultCnf.pData[0u] == 0);
 }
 
 int
